requirements.txt
transformers==4.46.3
torch==2.4.0
torchvision==0.19.0
python-dotenv==1.0.1
wandb==0.18.7